+++ Sources/Realtime/RealtimeChannel+AsyncAwait.swift
@@ -31,8 +31,7 @@ extension RealtimeChannelV2 {
     filter: String? = nil
   ) -> AsyncStream<InsertAction> {
     postgresChange(event: .insert, schema: schema, table: table, filter: filter)
-      .compactMap { $0.wrappedAction as? InsertAction }
-      .eraseToStream()
+      .compactErase()
   }
 
   /// Listen for postgres changes in a channel.
@@ -43,8 +42,7 @@ extension RealtimeChannelV2 {
     filter: String? = nil
   ) -> AsyncStream<UpdateAction> {
     postgresChange(event: .update, schema: schema, table: table, filter: filter)
-      .compactMap { $0.wrappedAction as? UpdateAction }
-      .eraseToStream()
+      .compactErase()
   }
 
   /// Listen for postgres changes in a channel.
@@ -55,8 +53,7 @@ extension RealtimeChannelV2 {
     filter: String? = nil
   ) -> AsyncStream<DeleteAction> {
     postgresChange(event: .delete, schema: schema, table: table, filter: filter)
-      .compactMap { $0.wrappedAction as? DeleteAction }
-      .eraseToStream()
+      .compactErase()
   }
 
   /// Listen for postgres changes in a channel.
@@ -104,10 +101,32 @@ extension RealtimeChannelV2 {
 
     return stream
   }
+  
+  /// Listen for `system` event.
+  public func system() -> AsyncStream<RealtimeMessageV2> {
+    let (stream, continuation) = AsyncStream<RealtimeMessageV2>.makeStream()
+
+    let subscription = onSystem {
+      continuation.yield($0)
+    }
+
+    continuation.onTermination = { _ in
+      subscription.cancel()
+    }
+
+    return stream
+  }
 
   /// Listen for broadcast messages sent by other clients within the same channel under a specific `event`.
   @available(*, deprecated, renamed: "broadcastStream(event:)")
   public func broadcast(event: String) -> AsyncStream<JSONObject> {
     broadcastStream(event: event)
+  }
+}
+
+// Helper to work around type ambiguity in macOS 13
+fileprivate extension AsyncStream<AnyAction> {
+  func compactErase<T: Sendable>() -> AsyncStream<T> {
+    AsyncStream<T>(compactMap { $0.wrappedAction as? T } as AsyncCompactMapSequence<Self, T>)
   }
 }