+++ Sources/Realtime/V2/CallbackManager.swift
@@ -1,10 +1,3 @@
-//
-//  CallbackManager.swift
-//
-//
-//  Created by Guilherme Souza on 24/12/23.
-//
-
 import ConcurrencyExtras
 import Foundation
 import Helpers
@@ -24,6 +17,10 @@ final class CallbackManager: Sendable {
 
   var callbacks: [RealtimeCallback] {
     mutableState.callbacks
+  }
+
+  deinit {
+    reset()
   }
 
   @discardableResult
@@ -71,6 +68,15 @@ final class CallbackManager: Sendable {
     mutableState.withValue {
       $0.id += 1
       $0.callbacks.append(.presence(PresenceCallback(id: $0.id, callback: callback)))
+      return $0.id
+    }
+  }
+
+  @discardableResult
+  func addSystemCallback(callback: @escaping @Sendable (RealtimeMessageV2) -> Void) -> Int {
+    mutableState.withValue {
+      $0.id += 1
+      $0.callbacks.append(.system(SystemCallback(id: $0.id, callback: callback)))
       return $0.id
     }
   }
@@ -145,6 +151,19 @@ final class CallbackManager: Sendable {
     }
   }
 
+  func triggerSystem(message: RealtimeMessageV2) {
+    let systemCallbacks = mutableState.callbacks.compactMap {
+      if case .system(let callback) = $0 {
+        return callback
+      }
+      return nil
+    }
+
+    for systemCallback in systemCallbacks {
+      systemCallback.callback(message)
+    }
+  }
+
   func reset() {
     mutableState.setValue(MutableState())
   }
@@ -167,16 +186,23 @@ struct PresenceCallback {
   var callback: @Sendable (any PresenceAction) -> Void
 }
 
+struct SystemCallback {
+  var id: Int
+  var callback: @Sendable (RealtimeMessageV2) -> Void
+}
+
 enum RealtimeCallback {
   case postgres(PostgresCallback)
   case broadcast(BroadcastCallback)
   case presence(PresenceCallback)
+  case system(SystemCallback)
 
   var id: Int {
     switch self {
     case let .postgres(callback): callback.id
     case let .broadcast(callback): callback.id
     case let .presence(callback): callback.id
+    case let .system(callback): callback.id
     }
   }
 }